reorg + expose new things
